# 8257837: Performance regression in heap byte buffer views