--- a/src/hotspot/share/oops/methodData.cpp
+++ b/src/hotspot/share/oops/methodData.cpp
@@ -1586,7 +1586,8 @@ bool MethodData::profile_unsafe(const methodHandle& m, int bci) {
   Bytecode_invoke inv(m , bci);
   if (inv.is_invokevirtual()) {
     if (inv.klass() == vmSymbols::jdk_internal_misc_Unsafe() ||
-        inv.klass() == vmSymbols::sun_misc_Unsafe()) {
+        inv.klass() == vmSymbols::sun_misc_Unsafe() ||
+        inv.klass() == vmSymbols::jdk_internal_misc_ScopedMemoryAccess()) {
       ResourceMark rm;
       char* name = inv.name()->as_C_string();
       if (!strncmp(name, "get", 3) || !strncmp(name, "put", 3)) {